Create the framework for the example code
This change creates the execution framework for the VDD2 example code and
provides a simple example.

VDD2 example classes have a void method with no parameters that is annotated
with @Example.  When the path to the class file is passed to RunExamples, it
finds that method and executes it.

Testing done:

Ran RunExamples against a class with an @Example method.  The current example
class, Basic, will not run correctly due to a runtime linking error.  This is
caused because the classpath in the voodoo jar file is not set and because the
dependent libs (selenium et al) are not otherwise on the classpath.  A fix
for that is forthcoming.