Add classpath to voodoo.jar and libs to deploy dir
This change makes it possible to run code out of the Voodoo jar file without
having to specify the classpath of every single lib that selenium depends on.
Of course, the jar file still isn't entirely portable - it needs to have a
lib directory with the dependencies next to it - but this is entirely
acceptable for now.

Testing done:

Ran RunExamples against an example that uses Voodoo.  It finally worked - a
browser window opened and then immediately closed, as expected.