CB-31: Add wait functionality
----------------------------------------------------------------------

Added waitForElement, which waits for an element to exist on the DOM
and returns it

Added waitForElementRemoved, which waits for an element to not exists
on the DOM.

----------------------------------------------------------------------

Added ability to wait for an element to be within the bounds of the
screen. This is needed because waitForVisible and waitForClickable
do not test if the specified element is within bounds. For example,
if a drop from above menu is clicked and we wait for an element on
the menu like so:

	waitForVisible(myElement, args...).click();

This will fail because the wait will finish while the element
has a negative y coordinate, causing an unclickable exception.
waitForOnScreen wait until the element is within bounds of the
screen. Thus,

	waitForOnScreen(myElement, args...).click();

should not fail in this case.

----------------------------------------------------------------------

Also added unit tests for the above.

Added local html test files for testings waits.

Since cases where elements were moving on an off screen, the unit
tests load local files to test from instead that provide consistent
and testable behaviour.

----------------------------------------------------------------------