CB-255 Polling Fails if Window doesn't Exist
Webdriver now waits for a new window to exist before throwing an out of
bounds error. Added a test to confirm this behaviour.

Also Updated the HTTPCompontents to the latest version. I has having
library issues after rebasing.

Also fixed (then ignored) a test that wasn't testing the correct thing.
Originally, it was testing for the wrong error message because url was
malformed. After fixing the url, the test properly gets a timeout error
instead. However, now it takes a full minute for the test to complete.
I ignored the test since it's still useful, as it should pass, but it
takes to long to run it every time.