Updated the HTTPCompontents Library to latest
Also fixed (then ignored) atest that wasn't testing the correct thing.
Originally, it was testing for the wrong error message because url was
malformed. After fixing the url, the test properly gets a timeout error
instead. However, now it takes a full minute for the test to complete.
I ignored the test since it's still useful, as it should pass, but it
takes to long to run it every time.